Enables error conversion of spans generated with Zipkin instrumentation
Zipkin's format of the "error" tag was designed to allow for cheaper
indexing and filtering for the common case where the error has a short
message description.

FWIW, we advised OpenTracing about their convention of a boolean
encouraging people to write "error"=false redundantly, and that having
to specify an error with two commands was less efficient both in process
and in indexing, but the leaders proceeded anyway.

Back to the point.. In Zipkin (and perhaps other non-opentracing)
instrumentation, presence of a tag named "error" (implicitly of type
string) means an error occurred. Other tags can also be present, but for
example when we are aggregating, you only need to filter on spans who
have any tag named "error".

It is my guess that Haystack intends to be useful beyond sites who
exclusively link to OpenTracing to produce tracing data, especially as
there's almost no logic in the pipeline that has anything to do with
OpenTracing. However, there is this one spot.. basically the way error
processing works it expects the exact boolean conventions which will
definitely not work with a lot of instrumentation out there.

The following change makes a very small adjustment which will allow this
to work with ad-hoc or non OpenTracing instrumentation. It seems an easy
portability win and keeps the metrics aggregation nicely decoupled in a
way that could work out of box with mechanical conversion upstream.